traceroute save in database

+++ code/src/data_models/hunter_models/hop_response_model.py
@@ -7,4 +7,4 @@ class HopResponseModel(BaseModel):
     id: int
     ip_address: str
     ttl: int
-    rtt_ms: int
+    rtt_ms: float
